compilation
Former-commit-id: 3ff7308a633abfb8c3c229f24e5df79b76092f6f

+++ trip/iwave++/core++/lib/iwstack.cc
@@ -49,7 +49,7 @@ namespace TSOpt{
 	oc_itstack.push_back(it);
       } 
       else { /* case 2 starts: */
-	if(ic_itstack.size() < max_ic_size) {
+	  if((int) ic_itstack.size() < max_ic_size) {
 	  /* case 2a: store dynamic fields into incore stack */
 	  
 	  /* allocate storage */
